Fixes overflow issue in DataReceptionHandler::NextEvent.
DataReceptionHandler::NextEvent gets slot number from Scheduler's
GetSlotNumber, which divides currentTicks by slot period (128000). This
slot value is multiplied by 128000 and subtracted from currentTicks to
arrive at ticks value until DataRx's wakeup.

Value returned from Scheduler's GetSlotNumber is a UINT32, whereas
subtraction is done with a UINT64 value. When currentTicks goes beyond 2^32,
the subtraction fails and a very large positive number is the result. This
cause DataRx to never happen.

This commit changes all UINT32 in Scheduler and DataRx to UINT64.

+++ MicroFrameworkPK_v4_3/DeviceCode/Include/Samraksh/MAC/OMAC/CMaxTimeSync.h
@@ -41,7 +41,7 @@ typedef class CMaxTimeSync : public EventHandler{
 	UINT32 m_localAverage;
 	CMaxTSyncState m_state;
 	UINT32 m_messagePeriod;
-	UINT32 m_lastSlotExecuted;
+	UINT64 m_lastSlotExecuted;
 
 	BOOL IsPacketScheduled;
 
@@ -52,10 +52,10 @@ public:
 	//UINT16 m_leader;
 
 	void Initialize(UINT8 radioID, UINT8 macID);
-	UINT64 NextEvent(UINT32 currentSlotNum);
-  	UINT16 NextEventinSlots(UINT32 currentSlotNum);
+	UINT64 NextEvent(UINT64 currentSlotNum);
+  	UINT16 NextEventinSlots(UINT64 currentSlotNum);
 	//UINT64 NextEvent(UINT64 currentTicks);
-	void ExecuteEvent(UINT32 currentSlotNum);
+	void ExecuteEvent(UINT64 currentSlotNum);
 	UINT8 ExecuteEventDone();
 	void PostExecuteEvent();
 	void SetWakeup(bool shldWakeup);
