+++ MicroFrameworkPK_v4_3/DeviceCode/Include/Samraksh/MAC/OMAC/DataReceptionHandler.h
@@ -25,14 +25,14 @@ class DataReceptionHandler: public EventHandler {
 
 public:
 	UINT16	m_nextSeed, m_mask; // m_nextSeed stores the next seed to be used in calculating the next wakeup slot and the m_mask is used as a mask in the pseduo random function
-	UINT32 m_nextwakeupSlot;//This variable stores the wakeup time in absolute slot number
+	UINT64 m_nextwakeupSlot;//This variable stores the wakeup time in absolute slot number
 	//UINT64 m_nextwakeupTicks; //This variable stores the absolute tick number to be used in the next
 	UINT32 m_seedUpdateIntervalinSlots;
 
 
 	void Initialize(UINT8 radioID, UINT8 macID);
 	UINT64 NextEvent();
-	void UpdateSeedandCalculateWakeupSlot(UINT32 &wakeupSlot, UINT16 &next_seed, const UINT16 &mask, const UINT32 &seedUpdateIntervalinSlots,  const UINT32 &currentSlotNum );
+	void UpdateSeedandCalculateWakeupSlot(UINT64 &wakeupSlot, UINT16 &next_seed, const UINT16 &mask, const UINT32 &seedUpdateIntervalinSlots,  const UINT64 &currentSlotNum );
  	void ExecuteEvent();
 
 	void PostExecuteEvent();
